+++ apps/ledger-live-desktop/src/renderer/screens/manager/AppsList/index.tsx
@@ -16,6 +16,7 @@ import Quit from "~/renderer/icons/Quit";
 import { Device } from "@ledgerhq/live-common/hw/actions/types";
 import AppList from "./AppsList";
 import DeviceStorage from "../DeviceStorage/index";
+import ProviderWarning from "../ProviderWarning";
 import AppDepsInstallModal from "./AppDepsInstallModal";
 import AppDepsUnInstallModal from "./AppDepsUnInstallModal";
 import ErrorModal from "~/renderer/modals/ErrorModal/index";
@@ -195,8 +196,8 @@ const AppsList = ({
           device={device}
           deviceName={deviceName}
           isIncomplete={isIncomplete}
-          firmware={firmware}
         />
+        <ProviderWarning />
         <AppList
           optimisticState={optimisticState}
           state={state}