+++ apps/ledger-live-desktop/src/renderer/screens/manager/DeviceStorage/EditDeviceName.tsx
@@ -52,10 +52,18 @@ const EditDeviceName: React.FC<Props> = ({
   const [name, setName] = useState<string>(deviceName.slice(0, maxDeviceName));
   const [completed, setCompleted] = useState<boolean>(false);
   const [error, setError] = useState<Error | undefined | null>(null);
+  const [actionError, setActionError] = useState<Error | undefined | null>(null);
   const [running, setRunning] = useState(false);
+  const [nonce, setNonce] = useState(0);
+
   const disableButton = running || (!completed && (deviceName === name || !name.trim() || error));
 
   const request = useMemo(() => ({ name }), [name]);
+  const onRetry = useCallback(() => {
+    setError(null);
+    setActionError(null);
+    setNonce(nonce => nonce + 1);
+  }, []);
 
   const onCloseDrawer = useCallback(() => {
     onClose?.();
@@ -96,6 +104,7 @@ const EditDeviceName: React.FC<Props> = ({
   return (
     <Flex
       flexDirection="column"
+      key={`${nonce}_editDeviceName`}
       rowGap={5}
       height="100%"
       overflowY="hidden"
@@ -103,7 +112,16 @@ const EditDeviceName: React.FC<Props> = ({
       flex={1}
       data-test-id="device-rename-container"
     >
-      <Flex flex={1} flexDirection="column" justifyContent="space-between" overflowY="hidden">
+      <Text alignSelf="center" variant="h5Inter" mb={3}>
+        {t("deviceRename.title", { productName })}
+      </Text>
+      <Flex
+        flex={1}
+        px={12}
+        flexDirection="column"
+        justifyContent="space-between"
+        overflowY="hidden"
+      >
         {completed ? (
           <Flex
             flex={1}
@@ -114,18 +132,11 @@ const EditDeviceName: React.FC<Props> = ({
           >
             <BoxedIcon
               Icon={Icons.CheckAloneMedium}
-              iconColor="success.c50"
+              iconColor="success.c60"
               size={64}
               iconSize={24}
             />
-            <Text
-              variant="large"
-              alignSelf="stretch"
-              mx={16}
-              mt={10}
-              textAlign="center"
-              fontSize={22}
-            >
+            <Text variant="large" alignSelf="stretch" mt={9} textAlign="center">
               <TextEllipsis>
                 {t("deviceRename.renamed", {
                   productName,
@@ -135,22 +146,18 @@ const EditDeviceName: React.FC<Props> = ({
             </Text>
           </Flex>
         ) : running ? (
-          <Flex flex={1} alignItems="center" justifyContent="center" p={2}>
+          <Flex flex={1} alignItems="center" justifyContent="center">
             <DeviceAction
               request={request}
               action={action}
+              inlineRetry={false}
               onResult={onSuccess}
-              // @ts-expect-error TODO: device and onClose are not expected as props - is it safe to remove?
-              device={device}
-              onClose={onClose}
+              onError={(error: Error) => setActionError(error)}
             />
           </Flex>
         ) : (
-          <Flex px={5} flexDirection="column">
-            <Text alignSelf="center" variant="h3Inter" mb={3}>
-              {t("deviceRename.title", { productName })}
-            </Text>
-            <Box flow={1} mb={5} p={5}>
+          <Flex flexDirection="column">
+            <Box flow={1} mb={5}>
               <Label mb={10} htmlFor="currentDeviceName">
                 {t("deviceRename.chooseName")}
               </Label>
@@ -165,10 +172,29 @@ const EditDeviceName: React.FC<Props> = ({
             </Box>
           </Flex>
         )}
-        {(!running || completed) && (
-          <Flex flexDirection="column" rowGap={8}>
-            <Divider />
-            <Flex alignSelf="end" px={12} pb={8}>
+      </Flex>
+      {!running || (running && actionError) || completed ? (
+        <Flex flexDirection="column" alignSelf="stretch">
+          <Divider />
+          <Flex
+            px={12}
+            alignSelf="stretch"
+            flexDirection="row"
+            justifyContent="space-between"
+            pt={4}
+            pb={1}
+          >
+            <Flex flex={1} />
+            {actionError ? (
+              <Button
+                data-test-id="retry-device-rename-button"
+                variant="main"
+                onClick={onRetry}
+                disabled={!running}
+              >
+                {t(`common.retry`)}
+              </Button>
+            ) : (
               <Button
                 variant="main"
                 onClick={completed ? onCloseDrawer : onSubmit}
@@ -176,10 +202,10 @@ const EditDeviceName: React.FC<Props> = ({
               >
                 {completed ? t(`common.close`) : t(`common.continue`)}
               </Button>
-            </Flex>
+            )}
           </Flex>
-        )}
-      </Flex>
+        </Flex>
+      ) : null}
     </Flex>
   );
 };