+++ libs/ledger-live-common/src/hw/actions/staxFetchImage.ts
@@ -1,6 +1,6 @@
 import { Observable } from "rxjs";
 import { scan, tap } from "rxjs/operators";
-import { useEffect, useState } from "react";
+import { useCallback, useEffect, useState } from "react";
 import { log } from "@ledgerhq/logs";
 import type { DeviceInfo } from "@ledgerhq/types-live";
 import { useReplaySubject } from "../../observable";
@@ -29,7 +29,11 @@ type State = {
   imageAlreadyBackedUp?: boolean;
 };
 
-type FetchImageAction = Action<FetchImageRequest, State, boolean>;
+type ActionState = State & {
+  onRetry: () => void;
+};
+
+type FetchImageAction = Action<FetchImageRequest, ActionState, boolean>;
 
 const mapResult = ({ completed, imageFetched, imageAlreadyBackedUp }: State) =>
   completed || imageFetched || imageAlreadyBackedUp || null;
@@ -73,6 +77,7 @@ const reducer = (state: State, e: Event): State => {
     case "appDetected":
       return {
         ...state,
+        error: null,
         unresponsive: false,
         requestQuitApp: true,
         isLoading: false,
@@ -80,6 +85,7 @@ const reducer = (state: State, e: Event): State => {
     case "imageFetched":
       return {
         ...state,
+        error: null,
         unresponsive: false,
         isLoading: false,
         fetchingImage: false,
@@ -89,6 +95,7 @@ const reducer = (state: State, e: Event): State => {
     case "currentImageHash":
       return {
         ...state,
+        error: null,
         unresponsive: false,
         isLoading: false,
         fetchingImage: true,
@@ -97,6 +104,7 @@ const reducer = (state: State, e: Event): State => {
     case "imageAlreadyBackedUp":
       return {
         ...state,
+        error: null,
         unresponsive: false,
         isLoading: false,
         fetchingImage: false,
@@ -107,6 +115,7 @@ const reducer = (state: State, e: Event): State => {
     case "progress":
       return {
         ...state,
+        error: null,
         unresponsive: false,
         isLoading: false,
         fetchingImage: true,
@@ -122,8 +131,9 @@ export const createAction = (
   const useHook = (
     device: Device | null | undefined,
     request: FetchImageRequest
-  ): State => {
+  ): ActionState => {
     const [state, setState] = useState(() => getInitialState(device));
+    const [resetIndex, setResetIndex] = useState(0);
     const deviceSubject = useReplaySubject(device);
 
     useEffect(() => {
@@ -153,10 +163,17 @@ export const createAction = (
       state.completed,
       state.imageAlreadyBackedUp,
       state.imageFetched,
+      resetIndex,
     ]);
+
+    const onRetry = useCallback(() => {
+      setResetIndex((currIndex) => currIndex + 1);
+      setState((s) => getInitialState(s.device));
+    }, []);
 
     return {
       ...state,
+      onRetry,
     };
   };