+++ tools/common-tools/config-overrides.js
@@ -32,6 +32,8 @@ module.exports = function override(webpackConfig) {
     "@ledgerhq/coin-framework/lib-es";
   webpackConfig.resolve.alias["@ledgerhq/coin-polkadot"] =
     "@ledgerhq/coin-polkadot/lib-es";
+  webpackConfig.resolve.alias["@ledgerhq/coin-algorand"] =
+    "@ledgerhq/coin-algorand/lib-es";
   webpackConfig.resolve.alias["@ledgerhq/domain-service"] =
     "@ledgerhq/domain-service/lib-es";